Support hashing dictionaries (maps)
I've chosen to implement maps as containing arbitrary values, rather
than a single type, i.e. value can be anything of type ObjectHashable.

Our use case is a bit simpler than this (values are always strings) but
I think it makes sense for the library to be a bit more flexible as
the original library is designed to support JSON-like structures.

However, I've decided to fix the keys as strings, because it makes
the implementation a bit simpler.